solve implicitly for anything except numbers
There is no reason to disallow solution of an equation for
a given object, so it is possible to isolate objects which are
not symbols as long as they are not numbers -- numbers should only
be themselves, so there should be nothing to solve for. Functions
like cos(1) are numbers and so are undefined functions like f(1)
but the latter are allowed since they don't evaluate to a number.