add note about implicit solving
Although solving for numbers 1, cos(1 + sqrt(2)) could be added,
this has not been added at this time. To do this, the user will
have to handle the substitution on their own.

Part of the reason this was avoided was because if expressions
are given that are themselves numbers, then trying to rearrange
for a number in that expression can only be done if the expression
itself is 0 -- if it's not then any inferences made from it are
invalid.  One has to watch for -1 that is hiding in all subtractions, too.

e.g. if eqs = Tuple(4 - sqrt(2), sqrt(2)/3 + 1)
>>> eqs = Tuple(4 - sqrt(2), sqrt(2)/3 + 1)
>>> eqs.subs([(1,a),(4,b)])
(-sqrt(2)*a + b, a + sqrt(2)/3)
>>> solve(_)
{b: -2/3, a: -sqrt(2)/3}

but compare to the solution after masing off the -1

>>> solve(eqs.subs(-1, Symbol('-1')).subs([(1,a),(4,b)]), a, b)
{b: -sqrt(2)*-1, a: sqrt(2)/(3*-1)}
>>> Dict(_).subs(Symbol('-1'), -1)
{a: -sqrt(2)/3, b: sqrt(2)}