2950: Rational processes Float and float
o The unprejudiced underlying fraction of the float is used to
create the Rational. Since as_integer_ratio() is not available in 2.5
(and Decimal in 2.5 will not take a float) mpf is used to parse the
float to get the mantissa and exponent.

o strings matching a literal 'float/float' are now processed and
processing is not passed to sympify (which makes Rational more literal).

        It makes sense to handle strings representing Rationals since
        it is about 100x faster than passing off the expression to
        sympify. No deeper processing is done, however.