restore simple multiset_partition counting
When a smart multiset_partitions generating routine is in place
it will be faster than what is here. I'm not aware of any other
way to solve this problem without actually generating the multisets.
Whatever work one does to avoid doing that seems to cost more than
simply generating the partitions of the multiset. The
one exception is in the counting of partitions of length 2 which can
be done very quickly. (There is perhaps some efficiency that can be
obtained for the case when there are only two different elements in the
multiset, e.g. aaabb, but for simplicity I have not included that.)