simplify bool mods
The code in simplify has been restored with a preceding
sympification (and quick exit in case expr is not Basic).

The code for `_eval_simplify` was moved from Boolean which
handles Eq (which is not a Boolean -- why is THAT happening??) to
BooleanFunction.

The tests in boolalg were patched with the correct return value
instead of the incorrect 'ans'.