stats: FooPSpace -> FooDensity
This is a large commit.

The information for the continuous distributions was pushed from a
pspace object into a density object.  Probability spaces are still the
base of the logic of stats, there are now just fewer of them.

Sampling had to move to the density level which brought along cdf and
inverse cdf.

Docstrings have not yet been updated