Travis-CI: add tests for Python 2.7/3.3 + gmpy
This adds two new tests:

Python 2.7 + gmpy
Python 3.3 + gmpy

At the moment, there are some test failures. So the tests are marked as
"Allowed Failures". Travis-CI then tests them, but keeps the results separate
from the main build. That way it will not disturb the SymPy workflow, but we
can work on fixing the failures, as well as manually examining the results to
make sure that no more failures are being introduced.