Change the colors in the Sphinx documentation
Basically, everything that used to be blue is now green, and everything that
used to be white (text) is now a light gray.  The theme is based off of the
SymPy website, SymPy Live, and SymPy Gamma.

I am not a web designer.  Probably all of this could be done a little bit
better.  Any suggestions are welcome.