Fixes ipython printing extension
This fixes two things.

First, the extension had to be reloaded to import the extension.
Fixes issue #3619

Second, this has been broken since IPython moved IPython.zmq to
IPython.kernel.zmq. This fixes this. Note, there are some git
versions of IPython where this will still break, but it should work
for all stable releases and will work for the current git version.