give evalf enough precision to compute number
I'm glad these tests failed. As far as I can see, they should not have
given the exact integer value because the default evalf was used. This
should only have given about 15 digits of precision and then the integer
part of the result would (and was) accurate only in the first 15 digits.
Now a precision of 70 is given which is enough to capture the full integer
part of the calculation.

I did not add tests of int without evalf but do note that one does
not need to evalf the expressions to a known precision in order to
obtain the correct integer part:

>>> int(floor(factorial(50)/E, evaluate=False))
11188719610782480504630258070757734324011354208865721592720336800L
>>> int(ceiling(factorial(50)/E, evaluate=False))
11188719610782480504630258070757734324011354208865721592720336801L