Use binding = prettyForm.MUL in several classes' printers
This makes the classes be parenthesized when used in a Pow.  This is an
alternate fix to #1788.  I changed Integral, Sum, Product, and Derivative. In
general, any kind of operator should use this, since it is printed like the
operator *times* the operand.

Also remove the special casing in print_Pow from #1788, since it was incorrect
in light of the existence of the binding settings.

This is a better fix of issue 3260.