A few edits to your edits...
added the asked-for description of f, lpart and pstack (at least starts for)

terminology: used components instead of elements, for multisets

Notes on count partitions: I meant in-order traversal (as opposed to
pre-order or post-order or breadth-first).

Thanks for the examples.

Style question: I had thought that the second line of a docstring was
supposed to be blank?  (E.g., your edit of multiset_partitions_taocp()
docstring -- am I out of date on this?)

	modified:   enumerative.py