fix geometry tests and modify arbitrary_point in line.py
Geometry's Line.arbitrary_point was changed so that it always maps
t=0 to the first point used to define the line and maps t=1 to the
second point. plot_interval was also changed so that Ray, like Line,
gives a Ray that is 10 units long.

Formerly, Ray had a definition that mapped t=0 to the origin of the
Ray and 1 to oo...but the ratio that did this, t/(1 - t), if simplified
to -t/(t - 1), mapped t=1 to -oo. Although a Piecewise could be used as

Piecewise(((oo, 1), (t/(1-t), True)))

defining it as described has the advantage of a unified parametric
definition of Linear entities.