radsimp modifications
Radsimp now handles the general solvable denominator that contains
4 radicals.

sign simplification is also employed on the return expression except
for the case where "sign trickery" is needed to keep the expression
from reverting to the original expression. It is perhaps a better idea
to use an unevaluated Mul with sorted args.

collect_sqrt is no longer used since that doesn't handle properly
the case for issue 2895.