heurisch_wrapper: Leave the original poles alone.
When integrating with heurisch_wrapper, we identify poles in the
denominator and reevaluate the integral at those poles. This is
pointless if the same poles where already in the integrand to
start with, so we now leave these poles alone.

Fixes issue 3827:
`integrate(sqrt(x**2/((y - x)*(y + x))), x)` showing some infinities
https://code.google.com/p/sympy/issues/detail?id=3827