Don't use unnamed Dummy objects in doctests
I changed _mask_nc to accept a name argument to allow this.  The doctest in
the Dummy docstring was skipped.

Resetting Dummy._count not only make the doctest confusing, but it's a bad idea
and could potentially lead to failures as some internal code could still have
Dummy objects laying around that would now compare equal to new, unrelated
ones.