rewrite hyperbolics unconditionally; inverse defined only for monotonics
inverse should be defined only if there is a single inverse for a given
value. sin(x) = 1/2 has 2 values per period and thus does not have an
inverse defined; tan(x) has a single value per period and has an inverse
defined. This attribute is used in the _invert function to algebraically
isolate the variable of interest. Functions with more than one value should
be handled in _tsolve or elsewhere.