remove dead code; annotate maybe-dead code
Also, after factoring, make sure that the expression
in _solve_lambert is fully inverted or else you can
end up with something like (x + 6*log(x))/2 = rhs --
the 2 should be combined with rhs. An existing test
will fail if this is not done. Alternatively the
code works fine if the 2 is distributed to give
x/2 + 3*log(x) so that is a possible thing to do, too.