Added slow test for mechanics.
Added a test which generates the equations of motion for a bicycle, linearizes
it, and compares the linear equations to benchmark values (numerically, using
eigenvalues). Most of the code was already in an example in the documentation,
just with the slow parts commented out - now it is actually a test and uses the
slow decorator.