Fabric: use absolute paths everywhere
See http://stackoverflow.com/q/17505712/161801. You can't assume that your cwd
is ~, because if the function is called from another function inside a cd()
context manager, it will be whatever that cd() set it to. The best thing to do
is to use absolute paths everywhere. Unfortunately, you can't even use ~,
because fabric treats those as relative paths. You have to spell out
/home/vagrant/