Make risch_integrate(x**x, x) work
Previously, this did not work, because it could not recognize x**x as
exp(x*log(x)) unless there already was a log(x) in the expression
somewhere. Now, we manually add log(x) to the extension as soon as we see a
nonelementary monomial of this form (with a symbolic base).

Of course, x**x has no elementary integral, but now risch_integrate is capable
of determining this without explicitly rewriting it as exp(x*log(x)). I don't
know of any examples of integrands that have elementary integrals of this form
with symbolic bases that don't also include the log of the base, but I haven't
looked very hard.

Also update the doctest in risch_integrate() to use this, because that's what
I wanted to use in the first place (I just never got around to fixing this).