Test Cases for coupled differntial system
Here i have just added the tests on which i have tried the correctness
and error in the code for cds. Tests are just to check the valid working
of cancel cases for primitive, hypertangent and hyperexponential
differntial extension. It also add the L_K and E_k to the DE,
 which would be required to suffice the conditions of the structure
theorem

The current code would seem to completely fail for hypertangent cases
as we have no support for T_K and A_K in prde.py
I have added the tests on which i have tried running the code
Though the outputs are not as expected(which tend to mostly fail because
of parametric log derivative problem returning None) but the code has been
tested interms of the implmentation wrt to cds.py
Improvments forth-coming commits