Fixed error in tests in mechanics.
Some tests in mechanics (or rather, various versions of one example problem)
had an error in the definition of the _kinematics_ of the rolling disc.
However, if you did the derivation by hand and also made this error (the
improper application of the 2-point theorem for acceleration for the rolling
w/o slip case) you would also get this error. That is the reason why the
original hand-calculated values matched the SymPy output. All is fixed now.

Additionally, the kinematic differential equations are calculated, rather than
being hard-coded. This should be more robust and less likely to be a source of
confusion/error in the future.