Fixed mech test failing?
One of the corrections to the tests seemed to fail on all the Python versions
except 3.2, which happens to be what I'm developing on. I've changed the
assertion from an equality test to a subtraction/expand/0-equality assertion -
I think this should make it more robust. I also tested it w/ python2.7,
successfully.