Fixed error with a slow test in mechanics.
The test, test_bicycle(), used to sort eigenvalues and compare the difference
(within a tolerance) to test for equality. For some reason, complex eigenvalues
stopped sorting. Now, the two matrices are directly compared, with the
difference in each element being compared (within a tolerance) and the minimum
value of the output matrix (which will be only booleans) needs to be True for
the test to pass.