Added dummify flag to lambdify.
Replaced logic in lambdify which attempted to determine if dummification should
or should not happen, with a flag (default True) for dummification. The idea is
if someone wants to read the output and are already supplying 'sympy' as a
module, they shouldn't have a problem supplying dummify=False as an argument.