Basic prototype of new ask using only satisfiable()
This is a basic prototype of a new ask, currently accessible as
sympy.assumptions.newask.newask(), that uses satisfiable() entirely to
determine the truth or falsehood of the given proposition under the given
assumptions.  See https://code.google.com/p/sympy/issues/detail?id=3929 for
more details on the idea.

As a first prototype, this is still in a very proof-of-concept stage. In
particular:

- I don't reuse any of the existing ask logic or machinery.  As such, almost
everything that works currently with ask() won't work with newask().

- I have so far only implemented the fact positive implies real.  Unless a
given proposition can be deduced from this fact, or is a tautology or a
contradiction, newask will return None.

- I have in no way attempted to make a nice API for this.  In fact, the
current API is rather terrible:  it's just a chain of if statements.

But the basic idea is there, namely, that ask() should extract all relevant
facts based on the predicates present in the proposition and the assumptions,
and then use satisfiable() to determine satisfiable(And(proposition,
assumptions, relevant_facts)) and satisfiable(And(~proposition, assumptions,
relevant_facts)).  If one is True and the other False, then we know that
proposition is True (or False); if both are True, we return None, and if both
are False, we conclude that the assumptions and the relevant_facts give a
contradiction.

At this point, I would like some input on how to structure this API so that it
becomes scalable, extensible, and readable.