Represent facts in newhandlers more symbolically
Instead of using complicated EquivalentAnyArgs and so on, just represent
symbolic things like (Mul, Equivalent(Q.zero, AnyArgs(Q.zero))) for the fact
that a multiplication is zero iff any term in it is zero.

Also rename handler registry to fact registry.

Fixed the fact about matrices (we only know a multiplication of matrices is
invertible iff each matrix is invertible if we know each matrix is
square). Now that I think about it, I'm not 100% sure if this is needed or not
because Q.invertible >> Q.square, but I'll leave it like this for now.

Add a fact to check if a number is prime (I haven't added any tests yet). This
shows how the UnevaluatedOnFree construction can be used to generate
computational facts, i.e., facts that compute their truth value from some
computation on the input (in this case, calling isprime() on an explicit
Integer).

What remains to do is to provide a better abstraction for representing facts
like Equivalent(Q.zero(x**y), Q.zero(x) & Q.positive(y)) (I hope that really
is equivalent). For now, we just represent these using a lambda expression,
but ideally, we would write exactly the above, and it would be pattern
matched.  This requires some thought, as it would be nice to continue to use
.rcall() uniformly everywhere.