_eval_interval: Use limit if simple subs resulted in oo.
_eval_interval relies on first a simple substitution, and then
falling back to the more correct limit calculation when we got NaN.
The problem is that it relies on things automatically simplifying
to NaN, but this is not always the case for complex expressions.
As soon as we get an Infinity, this should signal that we need
a more careful evaluation using the limit.