Make Predicates documentable
The issue is that we want each instance of Predicate to have its own
docstring. But Python's help() only looks at the __doc__ on classes.  The
solution is to trick it by creating a custom subclass of Predicate for each
instance, which takes on its instance's docstring.

Because each instance of Predicate has a custom subclass, code that previously
did things like "type(obj) == Predicate" or "obj.func == Predicate" (I'm
actually not 100% what .func should be) will not work, but code like that is
wrong anyway.  What you should do is isinstance(obj, Predicate), which still
works.

Also added tests to make sure that this logic all works, including with
subclasses.