4023: better handling of exponents in nfloat
When nfloat encountered RootOf instances the method
it used to leave the exponents alone didn't work because
RootOf instances don't allow non-univariate expression.
So RootOf(x**dummy_1 + x**dummy_2, 0) fails. The handling
of exponents is now handled in a more straightforward
manner and a test expression was added to solve's suite.